init split
