+++ web-console/src/components/rule-editor/rule-editor.scss
@@ -17,10 +17,6 @@
  */
 
 .rule-editor {
-  .right {
-    text-align: right;
-  }
-
   .title {
     display: flex;
