+++ web-console/src/console-application.tsx
@@ -23,26 +23,44 @@ import React from 'react';
 import type { RouteComponentProps } from 'react-router';
 import { Redirect } from 'react-router';
 import { HashRouter, Route, Switch } from 'react-router-dom';
+import type { Filter } from 'react-table';
 
 import type { HeaderActiveTab } from './components';
 import { HeaderBar, Loader } from './components';
 import type { DruidEngine, QueryWithContext } from './druid-models';
 import { Capabilities } from './helpers';
+import { stringToTableFilters, tableFiltersToString } from './react-table';
 import { AppToaster } from './singletons';
-import { localStorageGetJson, LocalStorageKeys, QueryManager } from './utils';
+import { compact, localStorageGetJson, LocalStorageKeys, QueryManager } from './utils';
 import {
   DatasourcesView,
   HomeView,
-  IngestionView,
   LoadDataView,
   LookupsView,
   SegmentsView,
   ServicesView,
   SqlDataLoaderView,
+  SupervisorsView,
+  TasksView,
   WorkbenchView,
 } from './views';
 
 import './console-application.scss';
+
+type FiltersRouteMatch = RouteComponentProps<{ filters?: string }>;
+
+function changeHashWithFilter(slug: string, filters: Filter[]) {
+  const filterString = tableFiltersToString(filters);
+  location.hash = slug + (filterString ? `/${filterString}` : '');
+}
+
+function viewFilterChange(slug: string) {
+  return (filters: Filter[]) => changeHashWithFilter(slug, filters);
+}
+
+function pathWithFilter(slug: string) {
+  return [`/${slug}/:filters`, `/${slug}`];
+}
 
 export interface ConsoleApplicationProps {
   defaultQueryContext?: Record<string, any>;
@@ -77,10 +95,8 @@ export class ConsoleApplication extends React.PureComponent<
 
   private supervisorId?: string;
   private taskId?: string;
-  private taskGroupId?: string;
-  private openDialog?: string;
-  private datasource?: string;
-  private onlyUnavailable?: boolean;
+  private openSupervisorDialog?: boolean;
+  private openTaskDialog?: boolean;
   private queryWithContext?: QueryWithContext;
 
   constructor(props: ConsoleApplicationProps) {
@@ -128,63 +144,76 @@ export class ConsoleApplication extends React.PureComponent<
   private resetInitialsWithDelay() {
     setTimeout(() => {
       this.taskId = undefined;
-      this.taskGroupId = undefined;
       this.supervisorId = undefined;
-      this.openDialog = undefined;
-      this.datasource = undefined;
-      this.onlyUnavailable = undefined;
+      this.openSupervisorDialog = undefined;
+      this.openTaskDialog = undefined;
       this.queryWithContext = undefined;
     }, 50);
   }
 
   private readonly goToStreamingDataLoader = (supervisorId?: string) => {
     if (supervisorId) this.supervisorId = supervisorId;
-    window.location.hash = 'streaming-data-loader';
+    location.hash = 'streaming-data-loader';
     this.resetInitialsWithDelay();
   };
 
   private readonly goToClassicBatchDataLoader = (taskId?: string) => {
     if (taskId) this.taskId = taskId;
-    window.location.hash = 'classic-batch-data-loader';
+    location.hash = 'classic-batch-data-loader';
     this.resetInitialsWithDelay();
   };
 
   private readonly goToDatasources = (datasource: string) => {
-    this.datasource = datasource;
-    window.location.hash = 'datasources';
-    this.resetInitialsWithDelay();
+    changeHashWithFilter('datasources', [{ id: 'datasource', value: `=${datasource}` }]);
   };
 
   private readonly goToSegments = (datasource: string, onlyUnavailable = false) => {
-    this.datasource = datasource;
-    this.onlyUnavailable = onlyUnavailable;
-    window.location.hash = 'segments';
+    changeHashWithFilter(
+      'segments',
+      compact([
+        { id: 'datasource', value: `=${datasource}` },
+        onlyUnavailable ? { id: 'is_available', value: '=false' } : undefined,
+      ]),
+    );
+  };
+
+  private readonly goToSupervisor = (supervisorId: string) => {
+    changeHashWithFilter('supervisors', [{ id: 'supervisor_id', value: `=${supervisorId}` }]);
+  };
+
+  private readonly goToTasksWithTaskId = (taskId: string) => {
+    changeHashWithFilter('tasks', [{ id: 'task_id', value: `=${taskId}` }]);
+  };
+
+  private readonly goToTasksWithTaskGroupId = (taskGroupId: string) => {
+    changeHashWithFilter('tasks', [{ id: 'group_id', value: `=${taskGroupId}` }]);
+  };
+
+  private readonly goToTasksWithDatasource = (datasource: string, type?: string) => {
+    changeHashWithFilter(
+      'tasks',
+      compact([
+        { id: 'datasource', value: `=${datasource}` },
+        type ? { id: 'type', value: `=${type}` } : undefined,
+      ]),
+    );
+  };
+
+  private readonly openSupervisorSubmit = () => {
+    this.openSupervisorDialog = true;
+    location.hash = 'supervisor';
     this.resetInitialsWithDelay();
   };
 
-  private readonly goToIngestionWithTaskId = (taskId?: string) => {
-    this.taskId = taskId;
-    window.location.hash = 'ingestion';
-    this.resetInitialsWithDelay();
-  };
-
-  private readonly goToIngestionWithTaskGroupId = (taskGroupId?: string, openDialog?: string) => {
-    this.taskGroupId = taskGroupId;
-    if (openDialog) this.openDialog = openDialog;
-    window.location.hash = 'ingestion';
-    this.resetInitialsWithDelay();
-  };
-
-  private readonly goToIngestionWithDatasource = (datasource?: string, openDialog?: string) => {
-    this.datasource = datasource;
-    if (openDialog) this.openDialog = openDialog;
-    window.location.hash = 'ingestion';
+  private readonly openTaskSubmit = () => {
+    this.openTaskDialog = true;
+    location.hash = 'tasks';
     this.resetInitialsWithDelay();
   };
 
   private readonly goToQuery = (queryWithContext: QueryWithContext) => {
     this.queryWithContext = queryWithContext;
-    window.location.hash = 'workbench';
+    location.hash = 'workbench';
     this.resetInitialsWithDelay();
   };
 
@@ -219,7 +248,10 @@ export class ConsoleApplication extends React.PureComponent<
         mode="all"
         initTaskId={this.taskId}
         initSupervisorId={this.supervisorId}
-        goToIngestion={this.goToIngestionWithTaskGroupId}
+        goToSupervisor={this.goToSupervisor}
+        goToTasks={this.goToTasksWithTaskGroupId}
+        openSupervisorSubmit={this.openSupervisorSubmit}
+        openTaskSubmit={this.openTaskSubmit}
       />,
       'narrow-pad',
     );
@@ -231,7 +263,10 @@ export class ConsoleApplication extends React.PureComponent<
       <LoadDataView
         mode="streaming"
         initSupervisorId={this.supervisorId}
-        goToIngestion={this.goToIngestionWithTaskGroupId}
+        goToSupervisor={this.goToSupervisor}
+        goToTasks={this.goToTasksWithTaskGroupId}
+        openSupervisorSubmit={this.openSupervisorSubmit}
+        openTaskSubmit={this.openTaskSubmit}
       />,
       'narrow-pad',
     );
@@ -243,13 +278,16 @@ export class ConsoleApplication extends React.PureComponent<
       <LoadDataView
         mode="batch"
         initTaskId={this.taskId}
-        goToIngestion={this.goToIngestionWithTaskGroupId}
+        goToSupervisor={this.goToSupervisor}
+        goToTasks={this.goToTasksWithTaskGroupId}
+        openSupervisorSubmit={this.openSupervisorSubmit}
+        openTaskSubmit={this.openTaskSubmit}
       />,
       'narrow-pad',
     );
   };
 
-  private readonly wrappedWorkbenchView = (p: RouteComponentProps<any>) => {
+  private readonly wrappedWorkbenchView = (p: RouteComponentProps<{ tabId?: string }>) => {
     const { defaultQueryContext, mandatoryQueryContext } = this.props;
     const { capabilities } = this.state;
 
@@ -267,14 +305,14 @@ export class ConsoleApplication extends React.PureComponent<
         capabilities={capabilities}
         tabId={p.match.params.tabId}
         onTabChange={newTabId => {
-          location.hash = `#workbench/${newTabId}`;
+          location.hash = `workbench/${newTabId}`;
         }}
         initQueryWithContext={this.queryWithContext}
         defaultQueryContext={defaultQueryContext}
         mandatoryQueryContext={mandatoryQueryContext}
         queryEngines={queryEngines}
         allowExplain
-        goToIngestion={this.goToIngestionWithTaskId}
+        goToTask={this.goToTasksWithTaskId}
       />,
       'thin',
     );
@@ -287,66 +325,93 @@ export class ConsoleApplication extends React.PureComponent<
       <SqlDataLoaderView
         capabilities={capabilities}
         goToQuery={this.goToQuery}
-        goToIngestion={this.goToIngestionWithTaskId}
+        goToTask={this.goToTasksWithTaskId}
       />,
     );
   };
 
-  private readonly wrappedDatasourcesView = () => {
+  private readonly wrappedDatasourcesView = (p: FiltersRouteMatch) => {
     const { capabilities } = this.state;
     return this.wrapInViewContainer(
       'datasources',
       <DatasourcesView
-        initDatasource={this.datasource}
+        filters={stringToTableFilters(p.match.params.filters)}
+        onFiltersChange={viewFilterChange('datasources')}
         goToQuery={this.goToQuery}
-        goToTask={this.goToIngestionWithDatasource}
+        goToTasks={this.goToTasksWithDatasource}
         goToSegments={this.goToSegments}
         capabilities={capabilities}
       />,
     );
   };
 
-  private readonly wrappedSegmentsView = () => {
+  private readonly wrappedSegmentsView = (p: FiltersRouteMatch) => {
     const { capabilities } = this.state;
     return this.wrapInViewContainer(
       'segments',
       <SegmentsView
-        datasource={this.datasource}
-        onlyUnavailable={this.onlyUnavailable}
+        filters={stringToTableFilters(p.match.params.filters)}
+        onFiltersChange={viewFilterChange('segments')}
         goToQuery={this.goToQuery}
         capabilities={capabilities}
       />,
     );
   };
 
-  private readonly wrappedIngestionView = () => {
+  private readonly wrappedSupervisorsView = (p: FiltersRouteMatch) => {
     const { capabilities } = this.state;
     return this.wrapInViewContainer(
-      'ingestion',
-      <IngestionView
-        taskId={this.taskId}
-        taskGroupId={this.taskGroupId}
-        datasourceId={this.datasource}
-        openDialog={this.openDialog}
+      'supervisors',
+      <SupervisorsView
+        filters={stringToTableFilters(p.match.params.filters)}
+        onFiltersChange={viewFilterChange('supervisors')}
+        openSupervisorDialog={this.openSupervisorDialog}
         goToDatasource={this.goToDatasources}
         goToQuery={this.goToQuery}
         goToStreamingDataLoader={this.goToStreamingDataLoader}
+        goToTasks={this.goToTasksWithDatasource}
+        capabilities={capabilities}
+      />,
+    );
+  };
+
+  private readonly wrappedTasksView = (p: FiltersRouteMatch) => {
+    const { capabilities } = this.state;
+    return this.wrapInViewContainer(
+      'tasks',
+      <TasksView
+        filters={stringToTableFilters(p.match.params.filters)}
+        onFiltersChange={viewFilterChange('tasks')}
+        openTaskDialog={this.openTaskDialog}
+        goToDatasource={this.goToDatasources}
+        goToQuery={this.goToQuery}
         goToClassicBatchDataLoader={this.goToClassicBatchDataLoader}
         capabilities={capabilities}
       />,
     );
   };
 
-  private readonly wrappedServicesView = () => {
+  private readonly wrappedServicesView = (p: FiltersRouteMatch) => {
     const { capabilities } = this.state;
     return this.wrapInViewContainer(
       'services',
-      <ServicesView goToQuery={this.goToQuery} capabilities={capabilities} />,
+      <ServicesView
+        filters={stringToTableFilters(p.match.params.filters)}
+        onFiltersChange={viewFilterChange('tasks')}
+        goToQuery={this.goToQuery}
+        capabilities={capabilities}
+      />,
     );
   };
 
-  private readonly wrappedLookupsView = () => {
-    return this.wrapInViewContainer('lookups', <LookupsView />);
+  private readonly wrappedLookupsView = (p: FiltersRouteMatch) => {
+    return this.wrapInViewContainer(
+      'lookups',
+      <LookupsView
+        filters={stringToTableFilters(p.match.params.filters)}
+        onFiltersChange={viewFilterChange('lookups')}
+      />,
+    );
   };
 
   render(): JSX.Element {
@@ -384,10 +449,15 @@ export class ConsoleApplication extends React.PureComponent<
                 <Route path="/sql-data-loader" component={this.wrappedSqlDataLoaderView} />
               )}
 
-              <Route path="/ingestion" component={this.wrappedIngestionView} />
-              <Route path="/datasources" component={this.wrappedDatasourcesView} />
-              <Route path="/segments" component={this.wrappedSegmentsView} />
-              <Route path="/services" component={this.wrappedServicesView} />
+              <Route path={pathWithFilter('supervisors')} component={this.wrappedSupervisorsView} />
+              <Route path={pathWithFilter('tasks')} component={this.wrappedTasksView} />
+              <Route path="/ingestion">
+                <Redirect to="/tasks" />
+              </Route>
+
+              <Route path={pathWithFilter('datasources')} component={this.wrappedDatasourcesView} />
+              <Route path={pathWithFilter('segments')} component={this.wrappedSegmentsView} />
+              <Route path={pathWithFilter('services')} component={this.wrappedServicesView} />
 
               <Route path="/query">
                 <Redirect to="/workbench" />
@@ -398,7 +468,7 @@ export class ConsoleApplication extends React.PureComponent<
               />
 
               {capabilities.hasCoordinatorAccess() && (
-                <Route path="/lookups" component={this.wrappedLookupsView} />
+                <Route path={pathWithFilter('lookups')} component={this.wrappedLookupsView} />
               )}
               <Route component={this.wrappedHomeView} />
             </Switch>
