+++ web-console/src/helpers/spec-conversion.ts
@@ -38,7 +38,7 @@ import type {
 import { inflateDimensionSpec, TIME_COLUMN, upgradeSpec } from '../druid-models';
 import { deepGet, filterMap, nonEmptyArray, oneOf } from '../utils';
 
-export function getSpecDatasourceName(spec: IngestionSpec): string {
+export function getSpecDatasourceName(spec: Partial<IngestionSpec>): string {
   return deepGet(spec, 'spec.dataSchema.dataSource') || 'unknown_datasource';
 }
