+++ web-console/src/views/home-view/supervisors-card/supervisors-card.tsx
@@ -68,8 +68,8 @@ FROM sys.supervisors`,
   return (
     <HomeViewCard
       className="supervisors-card"
-      href="#ingestion"
-      icon={IconNames.LIST_COLUMNS}
+      href="#supervisors"
+      icon={IconNames.ROUTE}
       title="Supervisors"
       loading={supervisorCountState.loading}
       error={supervisorCountState.error}
