+++ web-console/src/views/index.ts
@@ -18,10 +18,11 @@
 
 export * from './datasources-view/datasources-view';
 export * from './home-view/home-view';
-export * from './ingestion-view/ingestion-view';
 export * from './load-data-view/load-data-view';
 export * from './lookups-view/lookups-view';
 export * from './segments-view/segments-view';
 export * from './services-view/services-view';
 export * from './sql-data-loader-view/sql-data-loader-view';
+export * from './supervisors-view/supervisors-view';
+export * from './tasks-view/tasks-view';
 export * from './workbench-view/workbench-view';
