+++ web-console/src/views/load-data-view/load-data-view.spec.tsx
@@ -24,12 +24,28 @@ import { LoadDataView } from './load-data-view';
 
 describe('LoadDataView', () => {
   it('matches snapshot streaming', () => {
-    const loadDataView = shallow(<LoadDataView mode="streaming" goToIngestion={() => {}} />);
+    const loadDataView = shallow(
+      <LoadDataView
+        mode="streaming"
+        goToSupervisor={() => {}}
+        goToTasks={() => {}}
+        openSupervisorSubmit={() => {}}
+        openTaskSubmit={() => {}}
+      />,
+    );
     expect(loadDataView).toMatchSnapshot();
   });
 
   it('matches snapshot batch', () => {
-    const loadDataView = shallow(<LoadDataView mode="batch" goToIngestion={() => {}} />);
+    const loadDataView = shallow(
+      <LoadDataView
+        mode="batch"
+        goToSupervisor={() => {}}
+        goToTasks={() => {}}
+        openSupervisorSubmit={() => {}}
+        openTaskSubmit={() => {}}
+      />,
+    );
     expect(loadDataView).toMatchSnapshot();
   });
 });
