+++ web-console/src/views/load-data-view/load-data-view.tsx
@@ -127,6 +127,7 @@ import {
   updateSchemaWithSample,
   upgradeSpec,
 } from '../../druid-models';
+import { getSpecDatasourceName } from '../../helpers';
 import { getLink } from '../../links';
 import { Api, AppToaster, UrlBaser } from '../../singletons';
 import {
@@ -350,7 +351,10 @@ export interface LoadDataViewProps {
   mode: LoadDataViewMode;
   initSupervisorId?: string;
   initTaskId?: string;
-  goToIngestion: (taskGroupId: string | undefined, openDialog?: string) => void;
+  goToSupervisor: (supervisorId: string) => void;
+  openSupervisorSubmit: () => void;
+  goToTasks: (taskGroupId: string) => void;
+  openTaskSubmit: () => void;
 }
 
 interface SelectedIndex<T> {
@@ -570,6 +574,7 @@ export class LoadDataView extends React.PureComponent<LoadDataViewProps, LoadDat
     try {
       newSpec = upgradeSpec(newSpec);
     } catch (e) {
+      const streaming = isStreamingSpec(newSpec);
       newSpec = {};
       AppToaster.show({
         icon: IconNames.ERROR,
@@ -586,10 +591,8 @@ export class LoadDataView extends React.PureComponent<LoadDataViewProps, LoadDat
           </>
         ),
         action: {
-          text: 'Go to Ingestion view',
-          onClick: () => {
-            this.props.goToIngestion(undefined);
-          },
+          text: `Go to ${streaming ? 'Supervisors' : 'Tasks'} view`,
+          onClick: streaming ? this.props.openSupervisorSubmit : this.props.openTaskSubmit,
         },
       });
     }
@@ -1000,7 +1003,7 @@ export class LoadDataView extends React.PureComponent<LoadDataViewProps, LoadDat
   }
 
   renderWelcomeStepControls(): JSX.Element | undefined {
-    const { goToIngestion } = this.props;
+    const { openSupervisorSubmit, openTaskSubmit } = this.props;
     const { spec, selectedComboType } = this.state;
 
     const issue = this.selectedIngestionTypeIssue();
@@ -1089,7 +1092,7 @@ export class LoadDataView extends React.PureComponent<LoadDataViewProps, LoadDat
                 text="Submit supervisor"
                 rightIcon={IconNames.ARROW_RIGHT}
                 intent={Intent.PRIMARY}
-                onClick={() => goToIngestion(undefined, 'supervisor')}
+                onClick={openSupervisorSubmit}
               />
             </FormGroup>
             <FormGroup>
@@ -1097,7 +1100,7 @@ export class LoadDataView extends React.PureComponent<LoadDataViewProps, LoadDat
                 text="Submit task"
                 rightIcon={IconNames.ARROW_RIGHT}
                 intent={Intent.PRIMARY}
-                onClick={() => goToIngestion(undefined, 'task')}
+                onClick={openTaskSubmit}
               />
             </FormGroup>
           </>
@@ -3361,7 +3364,7 @@ export class LoadDataView extends React.PureComponent<LoadDataViewProps, LoadDat
   }
 
   private readonly handleSubmit = async () => {
-    const { goToIngestion } = this.props;
+    const { goToSupervisor, goToTasks } = this.props;
     const { spec, submitting } = this.state;
     if (submitting) return;
 
@@ -3384,7 +3387,7 @@ export class LoadDataView extends React.PureComponent<LoadDataViewProps, LoadDat
       });
 
       setTimeout(() => {
-        goToIngestion(undefined); // Can we get the supervisor ID here?
+        goToSupervisor(getSpecDatasourceName(spec));
       }, 1000);
     } else {
       let taskResp: any;
@@ -3405,7 +3408,7 @@ export class LoadDataView extends React.PureComponent<LoadDataViewProps, LoadDat
       });
 
       setTimeout(() => {
-        goToIngestion(taskResp.data.task);
+        goToTasks(taskResp.data.task);
       }, 1000);
     }
   };
