+++ web-console/src/views/services-view/services-view.spec.tsx
@@ -84,7 +84,14 @@ jest.mock('../../utils', () => {
 
 describe('ServicesView', () => {
   it('renders data', () => {
-    const comp = <ServicesView goToQuery={() => {}} capabilities={Capabilities.FULL} />;
+    const comp = (
+      <ServicesView
+        filters={[]}
+        onFiltersChange={() => {}}
+        goToQuery={() => {}}
+        capabilities={Capabilities.FULL}
+      />
+    );
 
     const servicesView = shallow(comp);
     expect(servicesView).toMatchSnapshot();
