+++ web-console/src/views/workbench-view/execution-details-pane/execution-details-pane.tsx
@@ -35,13 +35,13 @@ export type ExecutionDetailsTab = 'general' | 'sql' | 'native' | 'result' | 'err
 interface ExecutionDetailsPaneProps {
   execution: Execution;
   initTab?: ExecutionDetailsTab;
-  goToIngestion(taskId: string): void;
+  goToTask(taskId: string): void;
 }
 
 export const ExecutionDetailsPane = React.memo(function ExecutionDetailsPane(
   props: ExecutionDetailsPaneProps,
 ) {
-  const { execution, initTab, goToIngestion } = props;
+  const { execution, initTab, goToTask } = props;
   const [activeTab, setActiveTab] = useState<ExecutionDetailsTab>(initTab || 'general');
 
   function renderContent() {
@@ -59,7 +59,7 @@ export const ExecutionDetailsPane = React.memo(function ExecutionDetailsPane(
                 execution={execution}
                 onErrorClick={() => setActiveTab('error')}
                 onWarningClick={() => setActiveTab('warnings')}
-                goToIngestion={goToIngestion}
+                goToTask={goToTask}
               />
             ) : (
               <p>No stage info was reported.</p>
