+++ web-console/src/views/workbench-view/execution-progress-pane/execution-progress-pane.tsx
@@ -28,7 +28,7 @@ import './execution-progress-pane.scss';
 export interface ExecutionProgressPaneProps {
   execution: Execution;
   intermediateError?: Error;
-  goToIngestion(taskId: string): void;
+  goToTask(taskId: string): void;
   onCancel?(): void;
   allowLiveReportsPane?: boolean;
 }
@@ -36,7 +36,7 @@ export interface ExecutionProgressPaneProps {
 export const ExecutionProgressPane = React.memo(function ExecutionProgressPane(
   props: ExecutionProgressPaneProps,
 ) {
-  const { execution, intermediateError, onCancel, allowLiveReportsPane, goToIngestion } = props;
+  const { execution, intermediateError, onCancel, allowLiveReportsPane, goToTask } = props;
   const [showLiveReports, setShowLiveReports] = useState(true);
 
   return (
@@ -50,7 +50,7 @@ export const ExecutionProgressPane = React.memo(function ExecutionProgressPane(
         onCancel={onCancel}
       />
       {allowLiveReportsPane && showLiveReports && execution.stages && (
-        <ExecutionStagesPane execution={execution} goToIngestion={goToIngestion} />
+        <ExecutionStagesPane execution={execution} goToTask={goToTask} />
       )}
       {intermediateError && (
         <div className="network-error-notification">
