+++ web-console/src/views/workbench-view/execution-stages-pane/execution-stages-pane.spec.tsx
@@ -26,7 +26,7 @@ import { ExecutionStagesPane } from './execution-stages-pane';
 describe('ExecutionStagesPane', () => {
   it('matches snapshot', () => {
     const comp = shallow(
-      <ExecutionStagesPane execution={EXECUTION_INGEST_COMPLETE} goToIngestion={() => {}} />,
+      <ExecutionStagesPane execution={EXECUTION_INGEST_COMPLETE} goToTask={() => {}} />,
     );
 
     expect(comp).toMatchSnapshot();
