+++ web-console/src/views/workbench-view/query-tab/query-tab.tsx
@@ -81,7 +81,7 @@ export interface QueryTabProps {
   queryEngines: DruidEngine[];
   runMoreMenu: JSX.Element;
   clusterCapacity: number | undefined;
-  goToIngestion(taskId: string): void;
+  goToTask(taskId: string): void;
 }
 
 export const QueryTab = React.memo(function QueryTab(props: QueryTabProps) {
@@ -95,7 +95,7 @@ export const QueryTab = React.memo(function QueryTab(props: QueryTabProps) {
     queryEngines,
     runMoreMenu,
     clusterCapacity,
-    goToIngestion,
+    goToTask,
   } = props;
   const [alertElement, setAlertElement] = useState<JSX.Element | undefined>();
 
@@ -362,7 +362,7 @@ export const QueryTab = React.memo(function QueryTab(props: QueryTabProps) {
                 onDetails={onDetails}
                 queryEngines={queryEngines}
                 clusterCapacity={clusterCapacity}
-                goToIngestion={goToIngestion}
+                goToTask={goToTask}
               />
             ))}
             <div className={classNames('main-query', queryPrefixes.length ? 'multi' : 'single')}>
@@ -479,7 +479,7 @@ export const QueryTab = React.memo(function QueryTab(props: QueryTabProps) {
                     execution={execution}
                     onErrorClick={() => onDetails(statsTaskId!, 'error')}
                     onWarningClick={() => onDetails(statsTaskId!, 'warnings')}
-                    goToIngestion={goToIngestion}
+                    goToTask={goToTask}
                   />
                 )}
               </div>
@@ -501,7 +501,7 @@ export const QueryTab = React.memo(function QueryTab(props: QueryTabProps) {
               <ExecutionProgressPane
                 execution={executionState.intermediate}
                 intermediateError={executionState.intermediateError}
-                goToIngestion={goToIngestion}
+                goToTask={goToTask}
                 onCancel={onUserCancel}
                 allowLiveReportsPane
               />
